Populated UI
Added buttons for using weapons and potions as well as traveling between locations NESW of the player. Also added combo boxes for weapons or items the player can choose from later. There are also data views that will display a list of what the player has in their inventory, and what quests