Created use weapon and potion functions
Created and finished the use weapon and use potion functions so that upon using weapon it will calculate who is alive or not as well as how much damage to deal, then upon death what loot you receive. For the potion function users will use a potion and have their health updated accordingly, as well as their inventory. This SHOULD be a fully working game but I need to do some testing and create issues for any bugs